refactor: extract day notes hook
